Move props interface to types file
