+++ projects/js-packages/charts/src/components/leaderboard-chart/index.ts
@@ -1,4 +1,5 @@
 export { default as LeaderboardChart } from './leaderboard-chart';
-export type { LeaderboardChartProps, LeaderboardEntry } from './leaderboard-chart';
+export type { LeaderboardChartProps } from './types';
+export type { LeaderboardEntry } from '../../types';
 export { formatMetricValue } from '../../utils';
 export type { MetricValueType } from '../../utils';
