+++ projects/js-packages/charts/src/components/leaderboard-chart/leaderboard-chart.tsx
@@ -13,53 +13,10 @@ import {
 	useGlobalChartsContext,
 	useGlobalChartsTheme,
 } from '../../providers/chart-context';
-import { BaseChartProps, LeaderboardEntry } from '../../types';
+import { LeaderboardEntry } from '../../types';
 import { formatMetricValue } from '../../utils';
 import styles from './leaderboard-chart.module.scss';
-
-export interface LeaderboardChartProps extends BaseChartProps< LeaderboardEntry[] > {
-	/**
-	 * Whether to show comparison data
-	 */
-	withComparison?: boolean;
-
-	/**
-	 * Whether to overlay the label on top of bar
-	 */
-	withOverlayLabel?: boolean;
-
-	/**
-	 * Primary color for current period bars
-	 */
-	primaryColor?: string;
-
-	/**
-	 * Secondary color for comparison period bars
-	 */
-	secondaryColor?: string;
-
-	/**
-	 * Formatter for values
-	 */
-	valueFormatter?: ( value: number ) => string;
-
-	/**
-	 * Formatter for delta values
-	 */
-	deltaFormatter?: ( value: number ) => string;
-
-	/**
-	 * Whether the chart is in loading state
-	 */
-	loading?: boolean;
-
-	/**
-	 * Custom styling for the chart container
-	 */
-	style?: React.CSSProperties & {
-		'--a8c--charts--leaderboard--bar--border-radius'?: string;
-	};
-}
+import { LeaderboardChartProps } from './types';
 
 /**
  * Default value formatter using formatMetricValue
